requirements.txt
Brotli==1.1.0
Django==5.1.4
django-bootstrap4==24.4
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-database-url==1.0.3
django-js-asset==3.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
tzdata==2024.2
whitenoise==6.8.2
gunicorn==23.0.0
uvicorn==0.34.0